Refactor the installer for using some code in then upgrader
