+++ internal/pkg/edgectl/install/install.go
@@ -1,9 +1,7 @@
 package edgectl
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -22,7 +20,6 @@ import (
 	"github.com/spf13/cobra"
 	"helm.sh/helm/v3/pkg/strvals"
 	k8sTypesMetaV1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	k8sClientCoreV1 "k8s.io/client-go/kubernetes/typed/core/v1"
 	"k8s.io/client-go/rest"
 
@@ -261,17 +258,6 @@ func (i *Installer) CheckAESHealth() error {
 	return nil
 }
 
-// CheckHostnameFound tries to connect to check-blah.hostname to see whether DNS
-// has propagated. Each connect talks to a different hostname to try to avoid
-// NXDOMAIN caching.
-func (i *Installer) CheckHostnameFound() error {
-	conn, err := net.Dial("tcp", fmt.Sprintf("check-%d.%s:443", time.Now().Unix(), i.hostname))
-	if err == nil {
-		conn.Close()
-	}
-	return err
-}
-
 // FindMatchingHostResource returns a Host resource from the cluster that
 // matches the load balancer's address. The Host resource must refer to a
 // *.edgestack.me domain and be in the Ready state.
@@ -350,45 +336,6 @@ func (i *Installer) HostnameMatchesLBAddress(hostname string) bool {
 	return true
 }
 
-// CheckACMEIsDone queries the Host object and succeeds if its state is Ready.
-func (i *Installer) CheckACMEIsDone() error {
-
-	host, err := i.kubectl.Get("host", i.hostname, "")
-	if err != nil {
-		return LoopFailedError(err.Error())
-	}
-	state, _, err := unstructured.NestedString(host.Object, "status", "state")
-	if err != nil {
-		return LoopFailedError(err.Error())
-	}
-	if state == "Error" {
-		reason, _, err := unstructured.NestedString(host.Object, "status", "errorReason")
-		if err != nil {
-			return LoopFailedError(err.Error())
-		}
-		// This heuristic tries to detect whether the error is that the ACME
-		// provider got NXDOMAIN for the provided hostname. It specifically
-		// handles the error message returned by Let's Encrypt in Feb 2020, but
-		// it may cover others as well. The AES ACME controller retries much
-		// sooner if this heuristic is tripped, so we should continue to wait
-		// rather than giving up.
-		isAcmeNxDomain := strings.Contains(reason, "NXDOMAIN") || strings.Contains(reason, "urn:ietf:params:acme:error:dns")
-		if isAcmeNxDomain {
-			return errors.New("Waiting for NXDOMAIN retry")
-		}
-
-		// TODO: Windows incompatible, will not be bold but otherwise functions.
-		// TODO: rewrite Installer.show to make explicit calls to color.Bold.Printf(...) instead,
-		// TODO: along with logging.  Search for color.Bold to find usages.
-		i.ShowACMEFailed(reason)
-		return LoopFailedError(fmt.Sprintf("ACME failed. More information: kubectl get host %s -o yaml", i.hostname))
-	}
-	if state != "Ready" {
-		return errors.Errorf("Host state is %s, not Ready", state)
-	}
-	return nil
-}
-
 // CreateNamespace creates the namespace for installing AES
 func (i *Installer) CreateNamespace() error {
 	_ = i.kubectl.Create("namespace", defInstallNamespace, "")
@@ -399,6 +346,8 @@ func (i *Installer) CreateNamespace() error {
 
 // Perform is the main function for the installer
 func (i *Installer) Perform(kcontext string) Result {
+	var err error
+
 	chartValues := map[string]interface{}{}
 	for key, value := range defChartValues {
 		strvals.ParseInto(fmt.Sprintf("%s=%s", key, value), chartValues)
@@ -410,36 +359,10 @@ func (i *Installer) Perform(kcontext string) Result {
 	// Bold: Installing the Ambassador Edge Stack
 	i.ShowFirstInstalling()
 
-	// Attempt to grab a reasonable default for the user's email address
-	defaultEmail, err := i.Capture("get email", true, "", "git", "config", "--global", "user.email")
-	if err != nil {
-		i.log.Print(err)
-		defaultEmail = ""
-	} else {
-		defaultEmail = strings.TrimSpace(defaultEmail)
-		if !validEmailAddress.MatchString(defaultEmail) {
-			defaultEmail = ""
-		}
+	emailAddress, result := i.AskEmail()
+	if result.Err != nil {
+		return result
 	}
-
-	// Ask for the user's email address
-	i.ShowRequestEmail()
-
-	// Do the goroutine dance to let the user hit Ctrl-C at the email prompt
-	gotEmail := make(chan string)
-	var emailAddress string
-	go func() {
-		gotEmail <- getEmailAddress(defaultEmail, i.log)
-		close(gotEmail)
-	}()
-	select {
-	case emailAddress = <-gotEmail:
-		// Continue
-	case <-i.ctx.Done():
-		return i.resEmailRequestError(errors.New("Interrupted"))
-	}
-
-	i.log.Printf("Using email address %q", emailAddress)
 
 	// Beginning the AES Installation
 	i.ShowBeginAESInstallation()
@@ -460,7 +383,6 @@ func (i *Installer) Perform(kcontext string) Result {
 	}
 
 	i.restConfig, err = i.kubeinfo.GetRestConfig()
-
 	if err != nil {
 		return i.resGetRestConfigError(err)
 	}
@@ -657,90 +579,13 @@ func (i *Installer) Perform(kcontext string) Result {
 		}
 	}
 
-	i.ShowAESConfiguringTLS()
-
-	// Send a request to acquire a DNS name for this cluster's load balancer
-	regURL := "https://metriton.datawire.io/register-domain"
-	regData := &registration{Email: emailAddress}
-
-	if !metriton.IsDisabledByUser() {
-		regData.AESInstallId = i.clusterID
-		regData.EdgectlInstallId = i.scout.Reporter.InstallID()
-	}
-
-	if net.ParseIP(i.address) != nil {
-		regData.Ip = i.address
-	} else {
-		regData.Hostname = i.address
-	}
-
-	buf := new(bytes.Buffer)
-	_ = json.NewEncoder(buf).Encode(regData)
-	resp, err := http.Post(regURL, "application/json", buf)
-
-	if err != nil {
-		return i.resDNSNamePostError(err)
-	}
-
-	content, err := ioutil.ReadAll(resp.Body)
-	resp.Body.Close()
-
-	if err != nil {
-		return i.resDNSNameBodyError(err)
-	}
-
-	// With and without DNS.  In case of no DNS, different error messages and result handling.
-	dnsSuccess := true // Assume success with DNS
-	dnsMessage := ""   // Message for error reporting in case of no DNS
-	hostName := ""     // Login to this (hostname or IP address)
-
-	// Was there a DNS name post response?
-	if resp.StatusCode == 200 {
-		// Have DNS name--now wait for it to propagate.
-		i.hostname = string(content)
-		i.ShowAcquiringDNSName(i.hostname)
-
-		// Wait for DNS to propagate. This tries to avoid waiting for a ten
-		// minute error backoff if the ACME registration races ahead of the DNS
-		// name appearing for LetsEncrypt.
-
-		if err := i.loopUntil("DNS propagation to this host", i.CheckHostnameFound, lc2); err != nil {
-			return i.resDNSPropagationError(err)
-		}
-
-		i.Report("dns_name_propagated")
-
-		// Create a Host resource
-		hostResource := fmt.Sprintf(hostManifest, i.hostname, i.hostname, emailAddress)
-		if err := i.kubectl.Apply(hostResource, ""); err != nil {
-			return i.resHostResourceCreationError(err)
-		}
-
-		i.ShowObtainingTLSCertificate()
-
-		if err := i.loopUntil("TLS certificate acquisition", i.CheckACMEIsDone, lc5); err != nil {
-			return i.resCertificateProvisionError(err)
-		}
-
-		i.Report("cert_provisioned")
-		i.ShowTLSConfiguredSuccessfully()
-
-		if _, err := i.kubectl.Get("host", i.hostname, ""); err != nil {
-			return i.resHostRetrievalError(err)
-		}
-
-		// Made it through with DNS and TLS.  Set hostName to the DNS name that was given.
-		hostName = i.hostname
-		dnsSuccess = true
-	} else {
-		// Failure case: couldn't create DNS name.  Set hostName the IP address of the host.
-		hostName = i.address
-		dnsMessage = strings.TrimSpace(string(content))
-		i.ShowFailedToCreateDNSName(dnsMessage)
-		dnsSuccess = false
+	dnsMessage, hostName, resp, res := i.ConfigureTLS(emailAddress)
+	if res.Err != nil {
+		return result
 	}
 
 	// All done!
+	dnsSuccess := hostName == i.hostname
 	if dnsSuccess {
 		i.ShowAESInstallationComplete()
 	} else {
